feat: move portable shared components to @protolabs/ui molecules
Move hotkey-button, autocomplete, loading-state, error-state,
confirm-dialog, and markdown from apps/ui shared/ui to libs/ui
molecules. Update all imports across 30+ files to use
@protolabs/ui/molecules.
