+++ apps/ui/src/components/layout/sidebar/dialogs/trash-dialog.tsx
@@ -10,7 +10,7 @@ import {
 } from '@/components/ui/dialog';
 import { Button } from '@/components/ui/button';
 import { DeleteConfirmDialog } from '@/components/ui/delete-confirm-dialog';
-import { ConfirmDialog } from '@/components/ui/confirm-dialog';
+import { ConfirmDialog } from '@protolabs/ui/molecules';
 import type { TrashedProject } from '@/lib/electron';
 
 interface TrashDialogProps {
