+++ apps/ui/src/components/views/board-view/components/kanban-card/summary-dialog.tsx
@@ -10,7 +10,7 @@ import {
   DialogTitle,
 } from '@/components/ui/dialog';
 import { Button } from '@/components/ui/button';
-import { Markdown } from '@/components/ui/markdown';
+import { Markdown } from '@protolabs/ui/molecules';
 import { Sparkles } from 'lucide-react';
 
 interface SummaryDialogProps {
