+++ apps/ui/src/components/views/board-view/dialogs/follow-up-dialog.tsx
@@ -10,7 +10,7 @@ import {
 } from '@/components/ui/dialog';
 import { Button } from '@/components/ui/button';
 import { Label } from '@/components/ui/label';
-import { HotkeyButton } from '@/components/shared/hotkey-button';
+import { HotkeyButton } from '@protolabs/ui/molecules';
 import {
   DescriptionImageDropZone,
   FeatureImagePath as DescriptionImagePath,
