+++ apps/ui/src/components/views/github-issues-view.tsx
@@ -6,9 +6,9 @@ import { useQueryClient } from '@tanstack/react-query';
 import { getElectronAPI, GitHubIssue, IssueValidationResult } from '@/lib/electron';
 import { useAppStore } from '@/store/app-store';
 import { Button } from '@/components/ui/button';
-import { ConfirmDialog } from '@/components/ui/confirm-dialog';
-import { LoadingState } from '@/components/ui/loading-state';
-import { ErrorState } from '@/components/ui/error-state';
+import { ConfirmDialog } from '@protolabs/ui/molecules';
+import { LoadingState } from '@protolabs/ui/molecules';
+import { ErrorState } from '@protolabs/ui/molecules';
 import { cn, pathsEqual, generateUUID } from '@/lib/utils';
 import { toast } from 'sonner';
 import { queryKeys } from '@/lib/query-keys';
