+++ apps/ui/src/components/views/github-issues-view/components/comment-item.tsx
@@ -1,5 +1,5 @@
 import { User } from 'lucide-react';
-import { Markdown } from '@/components/ui/markdown';
+import { Markdown } from '@protolabs/ui/molecules';
 import type { GitHubComment } from '@/lib/electron';
 import { formatDate } from '../utils';
