+++ apps/ui/src/components/views/github-issues-view/dialogs/validation-dialog.tsx
@@ -7,7 +7,7 @@ import {
   DialogTitle,
 } from '@/components/ui/dialog';
 import { Button } from '@/components/ui/button';
-import { Markdown } from '@/components/ui/markdown';
+import { Markdown } from '@protolabs/ui/molecules';
 import {
   CheckCircle2,
   XCircle,
