+++ apps/ui/src/components/views/memory-view.tsx
@@ -31,7 +31,7 @@ import {
 import { Input } from '@/components/ui/input';
 import { Label } from '@/components/ui/label';
 import { cn } from '@/lib/utils';
-import { Markdown } from '../ui/markdown';
+import { Markdown } from '@protolabs/ui/molecules';
 import {
   DropdownMenu,
   DropdownMenuContent,
