+++ libs/ui/src/molecules/index.ts
@@ -1,1 +1,6 @@
-// Molecules will be populated by subsequent features
+export * from './hotkey-button.js';
+export * from './confirm-dialog.js';
+export * from './autocomplete.js';
+export * from './loading-state.js';
+export * from './error-state.js';
+export * from './markdown.js';
